requirements.txt
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
folium==0.12.1
geographiclib==1.50
geojson==2.5.0
geopy==2.1.0
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.3
overpass==0.7
overpy==0.6
psycopg2-binary==2.8.6
pygeoip==0.3.2
python-dotenv==0.17.1
requests==2.25.1
Shapely==1.7.1
urllib3==1.26.5
visitor==0.1.3
WTForms==2.3.3